+++ swid_generator/generators/softwareid_generator.py
